Moved Async and SameExecutorCompletionStage to org.jooq.lambda
